style(actionmenu): separate components to files; some style updates
